requirements.txt
annotated-types==0.7.0
anyio==4.9.0
atproto==0.0.61
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
db-dtypes==1.4.3
dnspython==2.7.0
emoji==2.14.1
fasttext==0.9.3
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
langdetect==1.0.9
libipld==3.1.1
numpy==2.3.1
oauthlib==3.3.1
packaging==25.0
pandas==2.3.0
pandas-gbq==0.29.1
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydata-google-auth==1.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
websockets==13.1
